requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-search-results==2.4.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic[email]==2.6.1
pytest==8.0.2
email-validator==2.1.0 